requirements.txt
absl-py==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Flask==2.2.3
future==0.18.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
Mako==1.2.4
MarkupSafe==2.1.2
oic==1.5.0
pycparser==2.21
pycryptodomex==3.17
pyjwkest==1.4.2
requests==2.28.2
six==1.16.0
solid_oidc_client==0.0.2
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
